feat: add grpc for user

+++ handler/v1/user/user.go
@@ -31,8 +31,8 @@ type LoginCredentials struct {
 
 // PhoneLoginCredentials 手机登录
 type PhoneLoginCredentials struct {
-	Phone      int `json:"phone" form:"phone" binding:"required" example:"13010002000"`
-	VerifyCode int `json:"verify_code" form:"verify_code" binding:"required" example:"120110"`
+	Phone      int64 `json:"phone" form:"phone" binding:"required" example:"13010002000"`
+	VerifyCode int   `json:"verify_code" form:"verify_code" binding:"required" example:"120110"`
 }
 
 // UpdateRequest 更新请求
