+++ internal/model/user_base.go
@@ -14,7 +14,7 @@ type UserBaseModel struct {
 	ID        uint64    `gorm:"primary_key;AUTO_INCREMENT;column:id" json:"id"`
 	Username  string    `json:"username" gorm:"column:username;not null" binding:"required" validate:"min=1,max=32"`
 	Password  string    `json:"password" gorm:"column:password;not null" binding:"required" validate:"min=5,max=128"`
-	Phone     int       `gorm:"column:phone" json:"phone"`
+	Phone     int64     `gorm:"column:phone" json:"phone"`
 	Email     string    `gorm:"column:email" json:"email"`
 	Avatar    string    `gorm:"column:avatar" json:"avatar"`
 	Sex       int       `gorm:"column:sex" json:"sex"`
